requirements.txt
Django>=3.0,<4.0
django-annoying
gevent
gunicorn
django-quilljs
Pillow
django-bootstrap4